вынес dom элементы в отдельный модуль, добавил блокировку слайдера при деактивации формы объявления
